ListAdapter implementation with ItemDiffCallback
