+++ app/src/main/java/com/leonid/myshoppinglist/presentation/MainActivity.kt
@@ -31,7 +31,7 @@ class MainActivity : AppCompatActivity() {
 
         viewModel = ViewModelProvider(this)[MainViewModel::class.java]
         viewModel.shopList.observe(this) {
-            shopListAdapter.shopList = it
+            shopListAdapter.submitList(it)
         }
     }
 
@@ -70,7 +70,7 @@ class MainActivity : AppCompatActivity() {
                 }
 
                 override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                    val shopItem = shopListAdapter.shopList[viewHolder.adapterPosition]
+                    val shopItem = shopListAdapter.currentList[viewHolder.adapterPosition]
                     viewModel.deleteShopItem(shopItem)
                 }
             }
