+++ app/src/main/java/com/leonid/myshoppinglist/presentation/ShopItemDiffCallback.kt
@@ -1,0 +1,15 @@
+package com.leonid.myshoppinglist.presentation
+
+import androidx.recyclerview.widget.DiffUtil
+import com.leonid.myshoppinglist.domain.ShopItem
+
+class ShopItemDiffCallback: DiffUtil.ItemCallback<ShopItem>() {
+
+    override fun areItemsTheSame(oldItem: ShopItem, newItem: ShopItem): Boolean {
+        return oldItem.id == newItem.id
+    }
+
+    override fun areContentsTheSame(oldItem: ShopItem, newItem: ShopItem): Boolean {
+        return oldItem == newItem
+    }
+}
