+++ app/src/main/java/com/leonid/myshoppinglist/presentation/ShopItemViewHolder.kt
@@ -1,0 +1,11 @@
+package com.leonid.myshoppinglist.presentation
+
+import android.view.View
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.leonid.myshoppinglist.R
+
+class ShopItemViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
+    val tvName: TextView = view.findViewById(R.id.tv_name)
+    val tvCount: TextView = view.findViewById(R.id.tv_count)
+}
