--- app/src/main/java/com/leonid/myshoppinglist/presentation/ShopListDiffCallback.kt
@@ -1,30 +1,0 @@
-package com.leonid.myshoppinglist.presentation
-
-import androidx.recyclerview.widget.DiffUtil
-import com.leonid.myshoppinglist.domain.ShopItem
-
-class ShopListDiffCallback(
-    private val oldList: List<ShopItem>,
-    private val newList: List<ShopItem>,
-): DiffUtil.Callback() {
-
-    override fun getOldListSize(): Int {
-        return oldList.size
-    }
-
-    override fun getNewListSize(): Int {
-        return newList.size
-    }
-
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        val oldItem = oldList[oldItemPosition]
-        val newItem = newList[newItemPosition]
-        return oldItem.id == newItem.id
-    }
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        val oldItem = oldList[oldItemPosition]
-        val newItem = newList[newItemPosition]
-        return oldItem == newItem
-    }
-}
